requirements.txt
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
dnspython==2.2.1
Flask==2.2.2
fonttools==4.37.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pymongo==4.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
gunicorn
python-dotenv